refactor(cpp): move platform to dedicated module
