+++ ios/SalveDbPlatform.mm
@@ -1,12 +1,12 @@
 #import <Foundation/Foundation.h>
-#include "../cpp/database/platform.hpp"
+#include "../cpp/platform/platform.hpp"
 
-namespace margelo::nitro::salvedb {
+namespace margelo::nitro::salvedb::platform {
 
 // s_baseDirectory is not used on iOS — documents dir is resolved on demand.
-void setPlatformDocumentsDirectory(const std::string&) {}
+void setDocumentsDirectory(const std::string&) {}
 
-std::string getPlatformDocumentsDirectory() {
+std::string getDocumentsDirectory() {
   NSArray<NSString*>* paths =
     NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES);
   if (paths.count == 0)
@@ -14,4 +14,4 @@ std::string getPlatformDocumentsDirectory() {
   return std::string(paths.firstObject.UTF8String);
 }
 
-} // namespace margelo::nitro::salvedb
+} // namespace margelo::nitro::salvedb::platform
